fix hm3
