+++ src/sass/page/login.scss
@@ -40,7 +40,7 @@
     margin-bottom: 30px;
 }
 
-.form__row-spacebetwin {
+.form__row_spacebetwin {
     justify-content: flex-end;
 }
