+++ src/templates/common/header.pug
@@ -14,14 +14,8 @@ header.header
                     svg.social__item-svg.social__item-svg_header
                         use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#in") 
     .header__hamburger.hamburger
-        button.cmn-toggle-switch.cmn-toggle-switch__htx
-            span toggle menu
-    //.header__hamburger.hamburger
-        .hamburger-box
-            .hamburger-inner
-        //button.hamburger__link
-            svg.hamburger__link-svg
-                use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#hamburger") 
+        button.hamburger__switch.hamburger__switch_htx
+            span hamburger
     .header__placeholder 
         svg.header__placeholder-title
             use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#portfolio_header")
