--- src/templates/common/menu.pug
@@ -1,10 +1,0 @@
-h2 Меню
-ul.menu
-    li
-        a(href="/index.html") Главная
-    li
-        a(href="/about.html") Обо мне
-    li
-        a(href="my-works.html") Мои работы
-    li
-        a(href="blog.html") Блог
