--- src/templates/common/technologies.pug
@@ -1,10 +1,0 @@
-p Я умею
-    ul.technologies
-        li
-            img(src='assets/images/sass.svg')
-        li
-            img(src='assets/images/pug.svg')
-        li
-            img(src='assets/images/gulp.svg')
-        li
-            img(src='assets/images/webpack.svg')
