+++ src/templates/pages/about.pug
@@ -5,16 +5,16 @@ block variables
 
 block content
     .wrapper
-        .menu
+        .full-menu
             ul.nav.nav_header
-                li.nav__item.nav__item_header
-                    a(href='my-works.html').nav__item-link.nav__item-link_header.block-title Мои работы
-                li.nav__item.nav__item_header
-                    a(href='about.html').nav__item-link.nav__item-link_header.block-title Обо мне
-                li.nav__item.nav__item_header
-                    a(href='blog.html').nav__item-link.nav__item-link_header.block-title Блог
-                li.nav__item.nav__item_header
-                    a(href='index.html').nav__item-link.nav__item-link_header.block-title Авторизация       
+                li.nav__item
+                    a(href='my-works.html').nav__item-link.block-title Мои работы
+                li.nav__item
+                    a(href='about.html').nav__item-link.block-title Обо мне
+                li.nav__item
+                    a(href='blog.html').nav__item-link.block-title Блог
+                li.nav__item
+                    a(href='index.html').nav__item-link.block-title Авторизация       
         include ../common/header.pug
         main.main
             section.section.about
@@ -26,21 +26,17 @@ block content
                                     use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#about_header")        
                             h1.block-title Обо мне
                             //.title-block__border    
-                    .about__left-content
+                    .about__left-content.user-description
                         .user-description__image
                             img(src="assets/images/content/face.jpg", alt="full phote").user-description__avatar
                         .user-description__title
-                            
                             h2.block-subtitle Кто я
-                                
                         .user-description__text.text-content
                             p Я супер веб разработчик из Новосибирска. Мне 27 лет. Я занимаюсь разработкой современных сайтов и приложений. Мне нравится делать интересные и современные проекты.
                             p Этот сайт я сделал в рамках обучения в Школе онлайн образования LoftSchool. Чуть позже я освежу в нём свой контент. А пока посмотрите,как тут всё классно и красиво!      
                 .about__right
                     .about__right-title
-                        
-                        h2.block-subtitle.block-subtitle_width Чем я могу быть вам полезен
-                            
+                        h2.block-subtitle.block-subtitle_width Чем я могу быть вам полезен  
                     .about__right-text.text-content
                         p Больше всего меня привлекет Frontend разработка, но я также знаком и могу решать не сложные задачи на Backend. Но давайте по порядку.
                     include ../common/skill.pug
@@ -49,7 +45,7 @@ block content
                 .map__contact
                     
                     h2.block-subtitle Контакты
-                    
+                
                     ul.contacts__list.contacts__list_margin
                         li.contacts__item
                             svg.contacts__item-icon
