+++ src/templates/pages/blog.pug
@@ -31,8 +31,8 @@ block content
                             svg.social__item-svg.social__item-svg_header
                                 use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#in") 
             .header__hamburger.hamburger
-                button.cmn-toggle-switch.cmn-toggle-switch__htx
-                    span toggle menu
+                button.hamburger__switch.hamburger__switch_htx
+                    span hamburger
             
             .header__placeholder.header__placeholder_blog 
                 svg.header__placeholder-title
@@ -56,11 +56,11 @@ block content
             section.section.blog
                 aside.blog__left
                     ul.blog-menu
-                        li.blog-menu__item.blog-menu__item-active
+                        li.blog-menu__item.blog-menu__item_active
                             button(href="#").blog-menu__item-link Самое важное в SASS
                         li.blog-menu__item 
                             button(href="#").blog-menu__item-link Самое важное в SASS
-                        li.blog-menu__item.blog-menu__item-active
+                        li.blog-menu__item.blog-menu__item_active
                             button(href="#").blog-menu__item-link Самое важное в SASS Самое важное в SASS Самое важное в SASS
                         li.blog-menu__item 
                             button(href="#").blog-menu__item-link Самое важное в SASS
