+++ src/templates/pages/index.pug
@@ -13,7 +13,7 @@ block content
           h1.visually-hidden Главная страница автора
           .welcome__content.welcome__content_margin
             .welcome__pic
-              img(src="assets/images/content/face.jpg", alt="Картинка автора").welcome__pic-img
+              img(src="assets/images/content/face.jpg").welcome__pic-img
             .welcome__description.description-info
               .description-info__name Владимир Астахов
               .description-info__about Личный сайт веб разработчика
@@ -30,7 +30,7 @@ block content
                 a.social__item-link(href='#')
                   svg.social__item-in.social__item-svg
                     use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#in") 
-          nav.nav
+          nav.nav.menu
             ul.menu__list
               li.menu__item
                 a.menu__item-link(href='my-works.html') Мои работы
@@ -43,7 +43,7 @@ block content
             form.form.form_margin(action="#")
               .form__row.form__row_margin-title
                   h2.block-subtitle.block-subtitle_white Авторизуйтесь
-              .form__row.form__row-spacebetwin.input-icon.input-icon_error
+              .form__row.form__row_spacebetwin.input-icon.input-icon_error
                 .input-icon__label
                   svg.input-icon__icon
                     use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#login") 
@@ -52,7 +52,7 @@ block content
                 .input-icon__error  
                   .input-icon__error-text Вы не ввели логин
 
-              .form__row.form__row-spacebetwin.input-icon.input-icon_error
+              .form__row.form__row_spacebetwin.input-icon.input-icon_error
                 .input-icon__label
                   svg.input-icon__icon
                     use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#password")  
@@ -62,7 +62,7 @@ block content
                   .input-icon__error-text Вы не ввели пароль
               .form__row   
                 label(for="checkbox").form__row-torggle 
-                  .visually-hidden Я человек
+                  span.visually-hidden Я человек
                   input(type="checkbox", name="checkbox" value="Я человек" id="checkbox").form__toggle-input
                   .form__toggle-fake
                     svg.form__toggle-fake_flag.form__toggle-fake_check
@@ -74,13 +74,13 @@ block content
                 p Вы точно не робот?
               .form__row.form__row_last
                 label(for="radio" ).form__row-torggle 
-                  .visually-hidden Да
+                  span.visually-hidden Да
                   input(type="radio", name="radio" value="Да" id="radio" checked).form__toggle-input
                   .form__toggle-fake.form__toggle-fake_radio
                     .form__toggle-fake_flag.form__toggle-fake_dot
                   .form__title Да  
                 label(for="radio_2").form__row-torggle 
-                  .visually-hidden Не уверен
+                  span.visually-hidden Не уверен
                   input(type="radio", name="radio" id="radio_2" value="Нет").form__toggle-input
                   .form__toggle-fake.form__toggle-fake_radio
                     .form__toggle-fake_flag.form__toggle-fake_dot
