+++ src/templates/pages/login.pug
@@ -10,13 +10,13 @@ block content
             form.form.form_margin(action="#")
               .form__row
                   h2.block-subtitle.block-subtitle_white Авторизуйтесь
-              .form__row.form__row-spacebetwin.input-icon
+              .form__row.form__row_spacebetwin.input-icon
                 .input-icon__label
                   svg.input-icon__icon
                     use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#login") 
                 label(for="login").visually-hidden Логин
                 input(type="text", name="login" placeholder="Логин" id="login").input-icon__login
-              .form__row.form__row-spacebetwin.input-icon
+              .form__row.form__row_spacebetwin.input-icon
                 .input-icon__label
                   svg.input-icon__icon
                     use(xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="assets/images/sprite/sprite.svg#password")
